pyproject.toml
[project]
name = "email-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "codebasetwo", email = "codebasetwo@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "google-api-python-client>=2.177.0",
    "google-auth-oauthlib>=1.2.2",
    "jupyter>=1.1.1",
    "langchain>=0.3.26",
    "langchain-chroma>=0.2.4",
    "langchain-cli>=0.0.36",
    "langchain-community>=0.3.27",
    "langchain-openai>=0.3.28",
    "langchain-tavily>=0.2.9",
    "langgraph>=0.5.3",
    "langgraph-cli[inmem]>=0.1.54",
    "langmem>=0.0.28",
    "openai>=1.97.0",
    "pip>=25.1.1",
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "python-dotenv>=1.1.1",
    "ruff>=0.12.4",
    "tavily-python>=0.7.9",
    "uvicorn>=0.35.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
